src/requirements.txt
django
django[bcrypt]
djangorestframework
djangorestframework-simplejwt[crypto]
django-cors-headers
psycopg[binary]